Fix SynchroSynth compilation
After the merge in commit e94b3c167ec the code of the `SynchroSynth`
class did not compile anymore due to several reasons. This commit
includes the fixes that were necessary to make it work again.

It's mostly:
* Removal of the `MM_OPERATORS` macro
* Switch from `sampleFrame` to the "new" `SampleFrame` class
* Adjusting some enums (`Plugin::Type::Instrument`, `Graph::Style::Linear`, `KnobType::Dark28`)
* Use `AudioEngine::outputSampleRate` instead of `AudioEngine::processingSampleRate`

The usage of `SampleFrame` also made slight changes with regards to the
assignment of the result samples necessary.

Fix warnings in three for loops which used `int` as the running variable
over an `unsigned int` type.