Make `EffectControlDialog` inherit `PluginView`
Before this commit, `EffectView` inherits `PluginView` and
`EffectControlDialog` inherits `ModelView`. This commit switches those
base classes.

The reason is that we previously had no base class for the 2 classes
`InstrumentView` and `EffectControlDialog`: `PluginView` was the base
class for `InstrumentView`, which does not contain the models for e.g.
the "Volume" knob, and for `EffectView`, which does contain models for
e.g. the "D/W" knob. This makes it questionable whether it is good that
`Effect` contains an `m_wetDryModel`, but `Instrument` contains no
"volume model" - but this commit does not fix this.

Notes:

1. This commit should be a no-op, but isn't exactly: While before,
   `EffectControlDialog` passed `EffectControls*` to the `ModelView`,
   now, it passes `Effect*` to the `PluginView`, which passes it
   directly to the `ModelView`.
2. The function `Effect::instantiateView(QWidget*)` seemed to be
   never executed. Nonetheless, it needs an implementation in order
   to provide a virtual method - we thus replace the content of this
   function with an `assert(false)`.