Use FileDialog::getSaveFileName
Use `FileDialog::getSaveFileName` to retrieve the file name of the file
into which the sample buffer is saved. This gives us lots of
functionality automatically, e.g. the request if an existing file should
be overwritten.

Also let remove the appending of the `.flac` file extension and let the
users decide using the system file dialog.