Use "summing" behavior rather than "averaging"
In order to support multiple inputs/outputs correctly, all track
channels routed to a single plugin input need to be summed together
rather than averaged, and likewise for all plugin outputs routed to a
single track channel.

This change in behavior is accompanied by a change to the default
connections for mono plugins so that it does not double the audio
amplitude. Only the left track (rather than both) are connected to the
plugin input in this case, matching what REAPER does.

I also updated the unit tests and added a new test specifically to test
the new "summing" behavior.