Fix rendering for Vectorscope
The rendering of the Vectorscope is broken on Wayland if the size of the
Vectorscope is increased. This is caused by using a `QImage` to render
the scope traces which is then scaled up.

Introduce a new way to paint the vector scope (goniometer) which simply
paints lines or points that progressively get dimmer and which does not
make use of a QImage anymore.

It supports the following features:
* Log mode
* Zooming
* Rendering the drawing performance
* Selecting a different color for the traces

It does not support:
* HQ Mode: The new implementation provides a performance that's
  equivalent to Non-HQ mode.
* Blurring of old data
* Persistence: Might be implemented by using a factor for the dimming

This commit introduces new controls:
* Lines mode: Switches between painting lines or points in the new
  implementation
* Legacy mode: Use to switch between the new and the old rendering. Only
  used during development as the old mode will be removed completely.

Rendering of the samples/trances uses the composition mode "Plus" so
that overlapping elements will appear like adding brightness. Painting
the grid and lines is done using the normal composition mode "Source
Over" so that it simply replaces existing pixels.

Painting of the lines/points and the grids and lines is done in a
"signal space", i.e. a transform where elements in the interval of
[-1, 1] feel "natural". The text is painted in "widget space".

The "HQ" button and the "Persistence" knob have been temporarily made
members so that they can be hidden in non-legacy mode. They will later
be removed completely.

Rendering in "Log" mode currently paints a horizontal line for silence.
It needs to be investigated why this is the case. It's likely caused by
the division by the distance of the points from the origin. If this
becomes very small (and non-zero) then the division might produce large
values.