Fix segfaults when deleting clips and tracks
If you simply erase the unique_ptr from the vector on removal, then when we reach the destructor of the derived track, the unique_ptr will be empty, and any attempt to read the array of tracks in the parent track container will result in a track being read that is nullptr. The same can be said for clips.

To fix this, we need to first make the unique_ptr release ownership of the object, remove the empty unique_ptr from the array, and then delete the object.